requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
holoviews
geoviews
seaborn
matplotlib
numpy
pandas
xgboost
sklearn
bokeh
statsmodels 
colorcet
cartopy
datashader 
geopandas